tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/non-animatable_properties.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - non-animatable properties in keyframes</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="Except animation-timing-function, properties
   that aren't animatable are ignored in keyframe rules">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes borderize {
         0% {
            border-right-width: 0px;
            border-right-style: none;
         }
         100% {
            border-right-width: 100px;
            border-right-style: none;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         border-right: solid 0px black;
      }
      .test {
         animation: borderize 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var passed;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var divStyle = getComputedStyle(div, null);
            var passed = (divStyle.borderRightWidth == '100px'
                          && divStyle.borderRightStyle == 'solid');
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>A black square should grow out from the right edge of the blue square
   below.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
